// src/app/shared/enums/role.ts
export enum Role {
    MANAGER = "giro-manager",

    MANAGER_PICKUP = "MANAGER_PICKUP",
    MANAGER_PRETREATMENT = "MANAGER_PRETREATMENT",
    MANAGER_IRA_OPERATION = "MANAGER_IRA_OPERATION",
    MANAGER_VALUER = "MANAGER_VALUER",
    MANAGER_THIRD_PARTY = "MANAGER_THIRD_PARTY",
    MANAGER_NOT_HIRED = "MANAGER_NOT_HIRED",
    MANAGER_HIRED_BY_BIDDING = "MANAGER_HIRED_BY_BIDDING",
    MANAGER_HIRED_BY_DIRECT_CONTRACT = "MANAGER_HIRED_BY_DIRECT_CONTRACT",
    MANAGER_HIRED_BY_BIDDING_AND_DIRECT_CONTRACT = "MANAGER_HIRED_BY_BIDDING_AND_DIRECT_CONTRACT",

    PRODUCER = "giro-producer",
    INDUSTRIAL_CONSUMER = "giro-industrial-consumer",
    CITY = "giro-city",

    ADMIN = "giro-admin",
    ADMIN_BIDDING = 'giro-admin-bidding',
    ADMIN_DOMICILIARY = 'giro-admin-domiciliary',
    ADMIN_NON_DOMICILIARY = 'giro-admin-non-domiciliary',
    ADMIN_PRODUCER = 'giro-admin-producer',

    USER_MEMBER = "USER_MEMBER",
    USER_MODERATOR = "USER_MODERATOR"
}