// src/app/shared/enums/hiring-status.ts
export enum HiringStatus {
    NOT_HIRED = "No contratado",
    THIRD_PARTY_MANAGER = "Gestor tercero",
    HIRED_BY_BIDDING = "Contratado por licitación",
    HIRED_BY_DIRECT_CONTRACT = "Contratado por contrato directo",
    HIRED_BY_BIDDING_AND_DIRECT_CONTRACT = "Contratado por licitación y contrato directo",
}